Split ConclusionResolver into ConclusionResolver and BoundConclusionResolver. Works except that it exposes a race condition when materialising.
